requirements.txt
ollama
PyMuPDF
gradio
langchain-community
langchain-chroma
groq
certifi
mistralai
langchain_groq
langchain-mistralai
mysql-connector-python
azure-identity